backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
Pillow
timm
python-multipart
opencv-python
numpy

pip freeze > requirements.txt